git-svn-id: https://appdevsvn.nj01/appdev/FreshDirect/features/branches/sf2_0_fe@80121 842219ff-9341-0410-a943-573460bdd4de

+++ projects/ERPSWebApp/src/com/freshdirect/webapp/action/fdstore/SubmitOrderAction.java
@@ -674,10 +674,10 @@ public class SubmitOrderAction extends WebActionSupport {
                 }
 							
 	    		//update newly added cartline ids
-	            Set<Integer> recentIds = new TreeSet<Integer>();
+	            Set<Long> recentIds = new TreeSet<Long>();
 	            for (FDCartLineI rc : modCart.getOrderLines()) {
 	            	if ( !(rc instanceof FDModifyCartLineI) ) {
-		                recentIds.add(Integer.valueOf(rc.getCartlineId()));
+		                recentIds.add(Long.valueOf(rc.getCartlineId()));
 	            	}
 	            }
 	    		user.setRecentCartlineIdsSet(orderNumber, recentIds);
